move bootstrap logic into a class
